requirements.txt
amqp==2.5.2
billiard==3.6.3.0
celery==4.4.1
Django==2.2
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
importlib-metadata==1.5.0
kombu==4.6.8
Markdown==3.2.1
Pillow==7.0.0
psycopg2-binary==2.8.4
PyJWT==1.7.1
python-memcached==1.59
pytz==2019.3
redis==3.4.1
six==1.14.0
sqlparse==0.3.1
vine==1.3.0
zipp==1.2.0
